Gun.cs added. PvP scene added. input for "Reload" needs to be added.
Gun.cs, the Projectile weapon code is added. There's also a Scene for PvP.
The only other thing needing to be done currently is sort out the input manager and get "Reload" added. What to bind it to will need to be discussed with the rest of the team. Most likely R.